Expose op_plan to python
A continuation of the cython sketch interface to OP2-Common.  The
planning functionality is wrapped up in an op_plan cython class.
To access the data needed for execution from the plan, use the class
methods that correspond to the OP2-Common struct slot name.  For
example, to get the loc_map slot from an op_plan instance, do:

   array = plan.loc_map()

This returns a 1-D numpy array of the correct type whose data buffer
points to the C side's data.  That means, DO NOT CHANGE the data you
get back from these calls.

Python code is responsible for marshalling the relevant bits of plan
data to the device if necessary.

There is no python object representing a plan, so lots of the glue
code is missing.