Implement vector map semantics in sequential backend
When an arg has a map but no index associated with it, we want all the
associated map indices.  The user kernel semantics for this are:

    void kernel(T *data[N])

where N is the dimension of the Dat.  To support this, when we see a
vector map argument we generate code (where M is the dimension of the
Map):

    T *data_vec[M];
    ...
    for i in range(set_size) {
        for idx in range(M) {
            data_vec[idx] = data + data_map[i * M + idx];
        }
        kernel(data_vec);
    }

Although these loops are unrolled in the generated code.

Since the vector map test now passes, remove the expected failure
marker from it.