Remove reliance on cython internals to extract op_mat pointer
Add a cptr property to the cython op_mat object.  This returns the
pointer to the C data structure as a uintptr_t (the integer type
guaranteed to be able to hold a void *).  In the generated code we now
pass this cptr property as an argument, convert it to an unsigned long
(aka uintptr_t) and then cast it to an op_mat.  This way, we don't
need to worry about how cython lays out its cdef classes.