Only populate backend dict with available backends
When we try and import a backend named 'foo' we get an ImportError if
it doesn't exist and can warn the user appropriately.  We no longer
need to maintain a list of implemented backends, we just try and load
the backend the user requested and raise an exception if that failed.
This means that an invalid backend now raises an ImportError rather
than a ValueError, so fix up the API test appropriately.