Fix _post_reduction_kernel_task code in opencl backend
The reduction_task_cache was keyed on the datatype, dimension and
operator of the reduction, however, the code itself also depended on
the name of the global.  Two globals with the same key, but different
names, would hit the cache and then fail to find the compiled code.
To fix this, make the name of the reduction kernel independent of the
name of the Global.