Introduce new function maybe_setflags to set flags on data arrays
The numpy documentation says of ndarray.setflags:

    WRITEABLE can only be set to True if the array owns its own
    memory or the ultimate owner of the memory exposes a writeable
    buffer interface or is a string.

In the case of data from an external application, PyOP2 neither owns
the data nor is that data a string.  We can therefore not call
setflags(write=True) on arrays wrapping this data (e.g. from
Fluidity).  Particularly unpleasantly, attempting such a setflags on
an array from Fluidity causes a segfault in numpy.

To work around these issues, only attempt to set the write flag on
data arrays owned by PyOP2.  This still gives the user a little
protection against shooting themselves in the foot.