Add free_sparsity routine
When a Sparsity object gets GC'd, we want to deallocate the large
arrays we've allocated for it with malloc.  If the data had been
allocated with the same mechanism as numpy uses, we could have just
set the OWNDATA flag on these arrays.  Unfortunately, there's no
guarantee that numpy uses malloc, so instead expose a free_sparsity
function which will free the data backing the arrays.