Fix transformation of user kernel in CUDA backend
The CUDA backend assumed that the user kernel function was the first
thing in the code string - this cannot be guaranteed to be the case with
FFC-generated kernels. This meant that functions weren't always getting
prepended with '__device__' correctly.

This commit fixes the issed by using pycparser to visit all the function
definitions and adding '__device__' to their specification.

Since the comment remover from opencl.py is required for this, it has
been factored into utils.py.