Get integer domain ID from new interface
UFL now stores the domain of a Measure object as a Region object.
A Region may have multiple subdomains. As a quick fix to get the
buildbot green again, we simply retrieve the first subdomain ID to
get the old behaviour, where only a single domain ID per measure
was supported (in all of UFL, FFC, and PyOP2). Full support for
Regions would require further work on PyOP2/ and FFC/PyOP2, and is
not a priority right now.