Outsource finding OP2 lib and include directories to find_op2 module
This module is imported by sequential and openmp and executed by
setup.py. That is necessary because setup.py cannot import a module
from a package without inheriting all its dependencies. Read more on
SO: http://stackoverflow.com/a/2073599/396967

Also support setting OP2_PREFIX instead of OP2_DIR to specify an
install location in a case where OP2 has not been built in the source
tree.