Add Vagrantfile and install script for automatic VM provisioning
This allows creating and launching a virtual machine with a single
command which comes with PyOP2 pre-installed. It requires VirtualBox
4.2 [1] and Vagrant [2] to be installed.

Simply run 'vagrant up' to automatically download the base VM image,
configure it for use with VirtualBox, boot the VM and install PyOP2
and all dependencies. The install script can also be used standalone
outside of the VM context.

[1]: https://www.virtualbox.org/wiki/Linux_Downloads
[2]: http://www.vagrantup.com/