base: Add name property to Arg object
This is constructed from the position of the Arg in the par_loop.  To
keep track of this, we record the position, and also the
"indirect_position" (the position of the first Arg that has the
dat/map pair of this Arg, this latter is necessary for generated
device code where we uniquify arguments passed to the stub).