Modify host code generation to correctly deal with MixedDat args
A MixedDat Arg is split when passed to the wrapper i.e. the data
pointers of the contained Dats are passed in. Therefore each Arg gets
an index into for the mixed component in the generated code, which
defaults to 0 for non-mixed Args.

We also need to use the right Map with the right arity when passing
an indirect Arg to the user kernel.