Remove mass2d_triangle and rename mass2d_unstructured regression test
The old mass2d_triangle regression test is removed since its "known
good" reference data is out of date and the mode of testing against
such a reference is not that valuable. So rather than updating the
reference, the test is removed and the old mass2d_unstructured test
is renamed to better reflect the demo it is testing.