DMPlex: Adding basic DMPlex interface
UnitSquare/Cube meshes are now built on-the-fly from a DMPlex object,
from which the cell_node_list is created when building a FunctionSpace.
For mesh files a simple DMPlex object is created from the old Fluidity
mesh object and handed over to the new DMPlex interface.

The new DMPlex interface is very basic and only works in serial with
non-extruded meshes. Higher order function spaces and boundary
conditions are also missing.