Only remove snes options when the solver is GC'd
This is necessary so that the options propagate into the sub-solves
correctly (by the time the options setting was called, the options the
user had passed in were no longer in the options database).

Since the snes object holds a reference to the solver it lives
inside (via the form_function/form_jacobian callbacks) we must destroy
it explicitly at the end of the solve_varproblem call.