Refactor configuration: Only create single instance
Instead of initialising op2.configuration to None and then setting it
to a Configuration object when op2.init is first called, set
base.configuration to the default configuration at import time and
then reconfigure when op2.init is called.

This means op2.configuration always points to the same object and
thereby issues with importing a configuration object from op2, which
is later re-assigned are avoided.