Build correct Sparsity for mixed measure forms
Previously, we assumed that bilinear forms would only assemble over
cells and hard-coded the Sparsity pattern to be the outer product of
the test and trial spaces' cell_node_maps.  For either pure facet
integrals, or mixed cell and facet integrals, this is the wrong thing
to do.  Instead, loop over the integrals in the form and construct the
appropriate map pairings to pass into Sparsity construction.