Fix matrix assembly for equal forms but different bcs
Previously, if the form we were assembling hit the matrix cache for an
already assembled matrix we would return that and not reset the boundary
conditions on it.  This was the wrong thing to do.  Equally, when
passing the matrix to assemble in as a tensor argument we would not
reset the boundary conditions either.  So do that.