Fix Mat._is_scalar_field for mixed x Non-mixed case
If one of datasets in a Mat is mixed and the other isn't, the dims
property is something like (x, (y, z)).  So calling np.prod when
determining if the Mat represents as scalar field fails.  Check for
scalar-ness by checking that np.prod of every entry in the tuple is 1
instead.