Use integrals from preprocessed form in assemble
If a form consists of multiple integrals of the same type, FFC returns
one kernel for each measure.  However, the unpreprocessed form may
return multiple integrals for the same measure, with the result that the
number of kernels does not match the number of integrals.  Instead, use
the integrals from the preprocessed form, where all measures have been
collapsed appropriately.  Any domain ids on the measure will now be
plain integers, so don't ask for their their subdomain_ids.