Correct error checking for Expression interpolation
Previously we were first checking whether the expression rank and
shape matches that of the function space. This was causing the wrong
error message for spaces that don't support point evaluation and
therefore one needs to use projection instead.

We initially make sure that we have an expression with a value for
each component in the value shape for each function space.