Implement BC application for extruded LHS assembly
In the extruded case, we can't fiddle with the maps directly to drop
matrix entries in assembly because that does the wrong thing for the
rest of the column.  Instead, add flags to the Map and iteration space
telling PyOP2 to generate code that fiddles with the map values
implicitly.