Use Map parent to compute colouring in plan
When imposing boundary conditions on matrices we make some of the map
values negative and remember where the map came from in the _parent
property.  Fix a segfault in plan construction when boundary conditions
are applied by always computing the colouring on the parent map if it
exists instead of the map itself.