Test assembly of volume integrals on surface of sphere
Integral of the cell area approximates the surface of the sphere.
Integral of an extruded annulus approximates the appropriate volume as
long as the layer height is very much less than the radius of the sphere
since we assume the Jacobian is independent of the position in the
element which is not actually the case.