Ensure space for diagonal entries in parallel matrix
In parallel, we don't allocate the complete sparsity pattern, only the
correct amount of space for non-zeros.  When we apply strong bcs, we
throw away entries, including on the diagonal, resulting in the final
assembly compressing that space away.  Explicitly place zeros on the
diagonal on first assembly so that there's always space for them.