DMPlex: Adding Plex-based extrusion
For extruded function spaces we use sections that point to stacked
vertical dof columns for each point in the base plex. When building
element->dof mappings we then traverse the closure explicitly and add
the DoFs belonging to the first element in the stack according to
explicit extrusion numbering rules.