Use split forms when assembling
Adapt to changes in compile_form, which now returns a four-tuple of
index, domain type, coefficients and kernel.

We now need to pull apart the tensor and test and trial functions if
it is a mixed type and extract the component for the block the kernel
contributes to and filter the boundary conditions to only include
those defined on the current block.