Simplify FormSplitter using splitting by construction
Instead of making one pass over the form and getting back a list of
subforms on the blocks of the mixed space, make as many passes as
there are blocks in the mixed space and only enabling one component of
each Argument at a time.

When splitting an Argument, create a UFL ListTensor where only the
component of the MixedFunctionSpace selected by the current block
index is included and all other components are set to Zero. These
Zeros are subsequently eliminated as they are used in expressions and
not propagated up the form tree.