Inject Firedrake root directory into sys.path for tests
This is required to be able to find the common module whichever
directory we are calling py.test from.

Note that this will ONLY work when tests are run by calling py.test,
not when calling them as a module. In that case it is required to have
the Firedrake root directory on your PYTYHONPATH to be able to call
tests from anywhere.