Implement our own CPU compilation
Instant is rather heavyweight for our needs, and it is very difficult to
modify either compilation flags or the compiler being used.  All we
really need to do is compile a single source file to a shared library
and load it with ctypes.  We can then set up the argument types to the
function we get back appropriately and call as before.