Implement Constant
Add a version of Constant (with data carried by a PyOP2 Global) that
behaves like an appropriate Coefficient to UFL and an appropriate
Function to Firedrake's assemble calls.

Effectively we build a FiniteElement, VectorElement, or TensorElement
for UFL and fake up enough Function methods that assembly needs no
changes to pass Constants in to the par_loop.