Add targets buildhtml and livehtml to docs Makefile
The former is required to avoid an infinite rebuild loop when the
livereload server is running: it is watch source to rebuild the docs
but also firedrake to rebuild apidoc. However the html make target
first invokes apidoc and thereby triggers an infinite loop. Hence add
a target buildhtml that does not depend on apidoc.