Rework function space caching to cache on the Mesh
Rather than holding on to every function space we ever build, cache
function spaces in the Mesh they're built on.  When the mesh goes out of
scope, its cache is no longer useful since we can't build any more
function spaces that will hit it.  This change stops memory leaks in the
case of building many meshes and function spaces in the same run.