Fix memory leak in JITModule on all backends
The JITModule is cached and therefore should not hold references to any
objects we might want to be garbage collected.  In particular, this
means it shouldn't hold on to the Args (since they hold references to
Dats and so forth) or, in the case of device backends, the ParLoop
itself.  To fix this, blow these slots away after use.