Add ufl_domain and ufl_id to Mesh
We build a ufl.Domain out of the coordinate field and use this to
construct finite element instances.  Because the coordinate field needs
a finite element, we perform a complicated bootstrapping procedure
whereby we first build a Domain on the Mesh's Cell with a data slot of
the Mesh itself (such that later on we can find the coordinates in
integrals) and use this Domain to build the finite element for the
coordinate field.  We then go back and replace the domain on the mesh
with a new one which is built on top the coordinate field.  Finally, we
replace the coordinate field itself with a new field sharing the same
data but this time built on a finite element with an actual domain (such
that it can be used a coefficient everywhere).