Accept kwargs in Expression construction
This allows users to define additional variables in the Expression
constructor that they can refer to by name in the code snippet.  Both
scalar and vector-valued variables are supported (the latter indexed in
the normal single-indirection way).  The value of the variable is
available on the Expression as a public property of the same name (which
also allows setting of values).