Implement fast derivation of P1 coarse cell to fine dof map
The map from coarse cells to fine dofs is isomorphic to the P2 map on
the coarse mesh.  This is significantly faster than the previous pure
python implementation, taking less than half a second to build a map
from 1280000 coarse cells to 2563201 fine dofs (in comparison, building
the coarse mesh in that case takes around 20 seconds).