Implement restriction of P1 fields on triangles
We restrict by computing a lumped projection of the fine field to the
coarse space.  To do this, we exploit that when the coarse cell is
pulled back to the reference cell, the fine cells always look the
same (since the refinement is uniform).  As a result, we can precompute
the coarse cell basis functions tabulated at the fine cell integration
points and then run over the mesh computing the RHS for the lumped
projection (u_f*phi_c*dx).