Support providing a form to act as preconditioner
Often we know something about our problem that means we can construct a
preconditioning matrix that is different from the operator itself.
Support this in the variational formulation solve interface by allowing
the user to pass an optional Jp form used to assemble the
preconditioning operator.  If Jp is not provided, we default to using
the provided operator to construct the preconditioner as before.  Note
for consistency in naming, when building a LinearVariationalProblem we
use aP (rather than Jp).