Reimplement top and bottom bcs for Mat assembly in extruded
Now that the Maps themselves keep track of whether any implicit boundary
conditions are to be applied, we can add back the code generation that
places negative values in the maps and do so on a map-by-map basis.
This means that we only apply bcs to the relevant part of a MixedMap.