Reimplement bcs for extruded spaces
In particular, rework implicit top and bottom bcs such that they only
apply to an individual Map, rather than the par_loop as a whole.  As a
side effect, assemble no longer needs to special case for these bc
types, since they're dealt with through the normal "flag the map values"
mechanism at a firedrake level.

As a result of this change, application of top and bottom boundary
conditions in mixed spaces now does the right thing and only applies the
bc to the relevant subspace.