Ignore decompression error when reading disk cache
Pickled objects cached on disk are gzip compressed. If this archive
is corrupted (because the write was interrupted or otherwise), the
file exists, but attempting to read it fails with a zlib.error
exception. We simply ignore this failure, which is subsequently
handled as if no value had been found in the cache.