Fully fill matrices with zeros on construction
The sparsity pattern we specify is the right one, but PETSc compresses
the allocation on matrix assembly to the number of "used" nonzeros.
This can result in requiring reallocation if (for example) the values
change because we're actually assembling the linearisation of a
non-linear operator.  To avoid this, when we first build a matrix, fill
all the entries we might ever use with zeros.