Save ParLoop objects on forms for repeated execution
If we see the same form with the same boundary conditions again when
assembling, we just replace the appropriate par_loop argument with the
new tensor and run the par_loop.

This reduces the "do-nothing" overhead for the following test:

    from firedrake import *
    parameters['assembly_cache']['enabled'] = False
    m = UnitTriangleMesh()
    V = FunctionSpace(m, 'DG', 0)
    u = TrialFunction(V)
    v = TestFunction(V)
    a = u*v*dx
    A = assemble(a)
    A.M
    %timeit assemble(a, tensor=A).M

from around 1.8 milliseconds to 180 microseconds (i.e. order of magnitude).