Fix GC issue for upstream UFL changes
Mesh objects are part of a reference cycle, since they hold on to a UFL
Domain which in turn references the Mesh.  But UFL objects have gained a
python-level __del__ which prevents collection of this cycle in
python2.  To work-around this, given that the default __del__ ob
ufl.Expr is a no-op, just remove it from the class definition after
importing UFL in firedrake.__init__.