Fix offset computation for exterior facet boundary map
Rather than building the offset by considering the global dofs in the
first entry of the exterior facet boundary node map and the cell to node
map (which was not guaranteed to coincide anyway), just index the offset
by the dofs FIAT has told us are on the boundary.  This not only
produces the correct offsets, but is robust to running in parallel where
a process may not own any dofs on the boundary (in which case indexing
the map would not work).