Build mesh hierarchies on non-overlapped meshes
So that the halos don't grow too large, we refine the non-overlapped
mesh and subsequently grow the halos on each level.  This allows us to
build reasonable multigrid hierarchies in parallel.  A consequence is
that we must postpone most of the mesh construction until we know we're
not going to build a mesh hierarchy.  This is transparent to the
user (the setup callback is called when building a function space or
similar).