Add tests of multigrid building blocks
These test restriction, prolongation and injection for a range of the
supported spaces.  Restriction is tested by restricting a simple
residual and confirming that this is equivalent to having integrated
that residual on a coarse mesh.  Prolongation is tested by choosing a
function exactly represented on the coarse mesh and ensuring that its
prolongation is equivalent to the same function interpolated onto the
fine mesh.  Injection is tested similarly, but in the opposite
direction.