Add tests of scaled mass matrix solves
We solve C*u*v*dx == C*v*dx expecting the answer 1.  C is chosen to be a
number of simple and complicated compound expressions involving
mathematical operators whose arguments are either a Function, or a
Constant, or a dot product of a vector-valued Function (or Constant)
with itself.  These forms are chosen to exercise various parts of the
COFFEE ast rewriter.