mg: Do not rely on internal PETSc details
Previously, we relied on implementation details of how the DMPlex halo
region was grown to determine maps between coarse and fine cells and
vertices.  This was fragile and broke with recent PETSc internal
changes.  To fix this state of affairs use public features to determine
the appropriate maps.  The idea is that the only canonical numbering
between the non-overlapped and overlapped DM object is the global
numbering.  So, to move between local points on one to local points on
the other we must first map from local to global point numbers using the
first DM's LGMap and subsequently map from global back to local numbers
using the second DM's LGMap.