Implement LRU expression cache
The generated kernels for pointwise expressions are cached on the
Function they're assigning into using an LRU cache to ensure it doesn't
get too large.  Additionally, to prevent excessive leakage of rvalues in
the expressions, the cache only holds weakrefs to the actual values (as
a result occasionally we hit the slow code path).

The exact speedup depends on the complexity of the expression being
assigned, but typical overhead reduction is 200ms, such that simple
assignments drop from ~300ms to 100ms.