install: Build h5py manually, not via pip
We need to link h5py against the same HDF5 library as PETSc.  By
default, the download script builds PETSc with libhdf5, so point h5py
there, rather than asking the package manager to download hdf5 and then
ignoring it in the PETSc build, but using it for h5py.