Track state changes in Mat assembly
When switching between inserting and adding values in a PETSc matrix, we
need to "flush" the assembly state.  We do this all the time when
assembling matrices with boundary conditions applied.  To facilitate
this switching, track the current state of a Mat object and make all
modification operations happen lazily.  When switching between insert
and add or vice versa we then know to call assembly flushing at the
correct point.

Modifications to the Mat (excepting from ParLoops) are now queued up
using a _LazyMatOp object which takes care of flushing assembly and
modifying the state of the matrix appropriately.