Add a test for another coffee issue.
This affects the jacobian calculation for the alpha term in the IP viscosity
for the full div(nu sym(grad(u))) viscosity form. With master coffee, this
currently fails for DG1 - but on the bendy branches this fails for the Hdiv
elements. This is all fixed in the fix-replacement branch of coffee.