travis: Enable build caching for PETSc and petsc4py
The firedrake-install script gains some travis-specific logic.  When
installing petsc or petsc4py, it checks to see if the travis cache is up
to date (the git sha matches).  Caches are invalidated and updated when
the git sha of the source code changes, but NOT if
PETSC_CONFIGURE_OPTIONS changes (a warning is in the code to this
effect).  Speeds up the firedrake-install phase of the travis build from
around 18 to 5 minutes.

Unfortunately, the OSX builder doesn't currently use the cache
infrastructure, however, to alleviate the possibility of timeouts during
firedrake-install (which can take a while!) we now print a
once-per-minute progress update in a background thread when running on
travis.  This should avoid travis terminating the job due to lack of
output.