Migrate FormSplitter to MultiFunction
Also take this opportunity to clean up the interface and document it.
Now we can always send all forms through the FormSplitter and it really
is a no-op in the non-mixed case.  Furthermore, remove usage of
sum_integrands, since this may well not be what we want.  The form
compiler should arrange to merge the integrals over the same measure.