Refactor code in assemble
There was a lot of duplication here, which we can simplify with some
judicious use of accessors.  At the same time, we remove the
timed_region decorators in the code.  When lazy evaluation is
on (basically all the time), these don't time the execution over the
mesh, but only the time to put the par_loop in the execution trace.  As
such, they almost always produce misleading information.