DMPlex: Add further reordering methods, including partitioner-based ones
In addition to `True` (RCM) and `False` (no reordering) the `reordering`
flag for mesh creation now also accepts strings `rcm`, `nd`, `1wd`, `qmd`
and `sfc`, as well as tuples of type `(name, nparts)` to indicate
partitioner-based reordering. For the latter the DMPlex's internal
partitioning is emulated; accepted values for name are `chaco` and
`parmetis`.